+++ .build/tasks/GuestConfig.build.ps1
@@ -40,7 +40,7 @@ param
 # SYNOPSIS: Building the Azure Policy Guest Configuration Packages
 task build_guestconfiguration_packages {
     # Get the vales for task variables, see https://github.com/gaelcolas/Sampler#task-variables.
-    . (Get-Command -Name 'Set-SamplerTaskVariable').ScriptBlock -AsNewBuild
+    . Set-SamplerTaskVariable -AsNewBuild
 
     if (-not (Split-Path -IsAbsolute $GCPackagesPath))
     {