+++ build.ps1
@@ -270,7 +270,7 @@ process
         }
 
         # Loading Build Tasks defined in the .build/ folder (will override the ones imported above if same task name).
-        Get-ChildItem -Path '.build/' -Recurse -Include '*.ps1' -ErrorAction Ignore |
+        Get-ChildItem -Path '.build/' -Recurse -Include '*.build.ps1' -ErrorAction Ignore |
             ForEach-Object {
                 "Importing file $($_.BaseName)" | Write-Verbose
 
@@ -313,7 +313,7 @@ process
     }
 }
 
-Begin
+begin
 {
     # Find build config if not specified.
     if (-not $BuildConfig)
@@ -503,4 +503,6 @@ Begin
 
         return
     }
+
+    Set-Alias -Name 'Set-SamplerTaskVariable' -Value "$PSScriptRoot/.build/tasks/Set-SamplerTaskVariable.ps1"
 }