+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Added a public function `Set-SamplerTaskVariable` that is used to dot-source
-  task variables for re-use over multiple build tasks.
+- Added a public alias `Set-SamplerTaskVariable` that points to the script
+  `Set-SamplerTaskVariable.ps1` in the tasks folder. This alias is used to
+  dot-source task variables for re-use over multiple build tasks.
 
 ### Fixed